requirements.txt
aiohttp=3.8.4
aiosignal=1.3.1
annotated-types=0.5.0
anyio=3.7.1
async-timeou=4.0.2
attrs=23.1.0
black=23.3.0
certifi=2023.5.7
charset-normalizer=3.2.0
click=8.1.4
dataclasses-json=0.5.9
exceptiongroup=1.1.2
fastapi=0.100.0
frozenlist=1.3.3
greenlet=2.0.2
h11=0.14.0
httptools=0.6.0
idna=3.4
langchain=0.0.228
langchainplus-sdk=0.0.20
marshmallow=3.19.0
marshmallow-enum=1.5.1
multidict=6.0.4
mypy-extensions=1.0.0
numexpr=2.8.4
numpy=1.25.1
openai=0.27.8
openapi-schema-pydantic=1.2.4
packaging=23.1
pathspec0.11.1
pip=23.1.2
platformdirs=3.8.1
pydantic=1.10.11
pydantic_core=2.1.2
python-dateutil=2.8.2
python-dotenv=1.0.0
python-environ=0.4.54
PyYAML=6.0
requests=2.31.0
setuptools=65.3.0
six=1.16.0
sniffio=1.3.0
SQLAlchemy=.0.18
starlette=0.27.0
tenacity=8.2.2
tomli=2.0.1
tqdm=4.65.0
typing_extensions=4.7.1
typing-inspect=0.9.0
urllib3=2.0.3
uvicorn=0.22.0
uvloop=0.17.0
watchfiles=0.19.0
websockets=11.0.3
yarl=1.9.2